+++ packages/commerce-sdk-react/src/auth/index.ts
@@ -364,6 +364,11 @@ class Auth {
         })
     }
 
+    clearUserAuth() {
+        this.logger.info('Login was invalidated. Clearing login state.')
+        this.clearStorage()
+    }
+
     /**
      * Every method in this class that returns a `TokenResponse` constructs it via this getter.
      */
@@ -951,6 +956,44 @@ class Auth {
         }
         this.clearStorage()
         return await this.loginGuestUser()
+    }
+
+    /**
+     * Handle updating customer password and re-log in after the access token is invalidated.
+     *
+     */
+    async updateCustomerPassword(body: {
+        customer: ShopperCustomersTypes.Customer
+        password: string
+        currentPassword: string
+    }) {
+        const {
+            customer: {customerId, login},
+            password,
+            currentPassword
+        } = body
+
+        // login and customerId are optional fields on the Customer type
+        // here we had to guard it to avoid ts error
+        if (!login || !customerId) {
+            throw new Error('Customer is missing required fields.')
+        }
+
+        const res = await this.shopperCustomersClient.updateCustomerPassword({
+            headers: {
+                authorization: `Bearer ${this.get('access_token')}`
+            },
+            parameters: {customerId},
+            body: {
+                password: password,
+                currentPassword: currentPassword
+            }
+        })
+        await this.loginRegisteredUserB2C({
+            username: login,
+            password
+        })
+        return res
     }
 
     /**