+++ packages/commerce-sdk-react/src/hooks/useMutation.test.ts
@@ -11,9 +11,11 @@ import {
     mockMutationEndpoints,
     renderHookWithProviders,
     waitAndExpectSuccess,
+    waitAndExpectError,
     DEFAULT_TEST_CONFIG
 } from '../test-utils'
-import {useCustomMutation} from './useMutation'
+import {useMutation, useCustomMutation} from './useMutation'
+import Auth from '../auth'
 
 jest.mock('../auth/index.ts', () => {
     const {default: mockAuth} = jest.requireActual('../auth/index.ts')
@@ -145,5 +147,72 @@ describe('useCustomMutation', () => {
         act(() => result.current.mutate({headers: mockHeaders}))
         await waitAndExpectSuccess(() => result.current)
         expect(result.current.data).toHaveProperty('test')
+    })
+    test('clear auth state when request uses invalid session', async () => {
+        const spy = jest.spyOn(Auth.prototype, 'clearUserAuth')
+        const mockRes = {
+            title: 'Unauthorized',
+            type: 'https://api.commercecloud.salesforce.com/documentation/error/v1/errors/unauthorized',
+            detail: 'Customer credentials changed after token was issued.'
+        }
+        const apiName = 'hello-world'
+        mockMutationEndpoints(apiName, mockRes, 401)
+
+        const {result} = renderHookWithProviders(() => {
+            return useCustomMutation({
+                options: {
+                    method: 'POST',
+                    customApiPathParameters: {
+                        endpointPath: 'test-hello-world',
+                        apiName
+                    },
+                    body: {test: '123'}
+                },
+                rawResponse: false
+            })
+        })
+
+        expect(result.current.error).toBeNull()
+        act(() => result.current.mutate())
+        await waitAndExpectError(() => result.current)
+        expect(spy).toHaveBeenCalled()
+    })
+})
+
+describe('useMutation', () => {
+    test('clear auth state when request uses invalid session', async () => {
+        const spy = jest.spyOn(Auth.prototype, 'clearUserAuth')
+
+        const mockRes = {
+            title: 'Unauthorized',
+            type: 'https://api.commercecloud.salesforce.com/documentation/error/v1/errors/unauthorized',
+            detail: 'Customer credentials changed after token was issued.'
+        }
+
+        const hookConfig = {
+            client: {},
+            method: () =>
+                new Promise((resolve, reject) => {
+                    reject({
+                        response: {
+                            json: () => mockRes
+                        }
+                    })
+                }),
+            getCacheUpdates: () => {}
+        }
+
+        const {result} = renderHookWithProviders(() => {
+            // set hookConfig as any since we're just trying to invoke the method and not an actual query
+            return useMutation(hookConfig as any)
+        })
+
+        act(() => result.current.mutate({}))
+
+        await waitAndExpectError(() => {
+            return result.current
+        })
+
+        expect(spy).toHaveBeenCalled()
     })
 })